Basics laufen

+++ StonksConsole/StockPriceData.cs
@@ -5,12 +5,12 @@ namespace StonksConsole
     public class StockPriceData
     {
         public DateTime Timestamp { get; set; }
-        public decimal Open { get; set; }
+        public float Open { get; set; }
 
-        public decimal High { get; set; }
-        public decimal Low { get; set; }
+        public float High { get; set; }
+        public float Low { get; set; }
 
-        public decimal Close { get; set; }
-        public decimal Volume { get; set; }
+        public float Close { get; set; }
+        public float Volume { get; set; }
     }
 }
